wsgi/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
python-redmine==2.4.0
mattermostautodriver==1.3.0
gunicorn==21.2.0
pytest==7.4.2
